// src/app/interfaces/song.ts
export interface Song {
  song_id: number;
  song_uid: string;
  song_title: string;
  description: string;
  created_on: string;
}

// export interface Song {
//   user_id: string;
//   song_id: number;
//   song_uid: string;
//   is_original: string;
//   sole_creator: string;
//   creative_status: string;
//   song_title: string;
//   artist_name: string;
//   genres: string;
//   moods: string;
//   instruments: string;
//   bpm: string;
//   lyrics: string;
//   description: string;
//   created_on: string;
// }